Fixs after oral presentation

+++ ui/groupe.js
@@ -5,11 +5,11 @@ const groupe = {
         <div id="paramGroupe">
             <div>
                 <span>Modifier le nom :</span>
-                <input class="Champ" id="Nom" v-model="groupe.nom">
+                <input class="Champ" id="Nom" v-model="groupe.nom" :readonly="groupe.id === 1">
             </div>
             <div>
                 <span id="span2">Couleur du groupe :</span>
-                <input type="color"class="Champ" id="Couleur" v-model="groupe.couleur">
+                <input type="color" class="Champ" id="Couleur" v-model="groupe.couleur">
             </div>
         </div>
 
@@ -18,18 +18,19 @@ const groupe = {
             <li v-for="contact in contactsDuGroupe" :key="contact.id" class="contact-item" @click="goToContactProfile(contact.id)">
                 <p>{{ contact.nom }} {{ contact.prenom }}</p>
             </li>
-            <div class="buttons">
-                <button id="But1" @click="goBack">Retour</button>
-                <button id="But2" @click="confirmDelete">Supprimer</button>
-                <button id="But3" @click="saveGroupe">Enregistrer</button>
-            </div>
+        </div>
+        <div class="buttons">
+            <button id="But1" @click="goBack">Retour</button>
+            <button id="But2" @click="confirmDelete">Supprimer</button>
+            <button id="But3" @click="saveGroupe">Enregistrer</button>
         </div>
     </div>
     `,
     data() {
         return {
             groupe: null,
-            contactsDuGroupe: []
+            contactsDuGroupe: [],
+            originalNom: '' // Pour stocker le nom original lors du chargement
         };
     },
     methods: {
@@ -37,7 +38,6 @@ const groupe = {
             this.$router.push('/groupesPage');
         },
         goToContactProfile(contactId) {
-            // Rediriger vers la page de profil du contact
             this.$router.push(`/contact/${contactId}`);
         },
         loadContactsDuGroupe() {
@@ -53,6 +53,11 @@ const groupe = {
                 .catch(error => console.error('Erreur lors du chargement des contacts du groupe:', error));
         },
         saveGroupe() {
+            if (this.groupe.id === 1 && this.groupe.nom !== this.originalNom) {
+                alert('Le nom du groupe Favoris ne peut pas être modifié.');
+                this.groupe.nom = this.originalNom; // Réinitialiser le nom à 'Favoris'
+                return;
+            }
             axios.put(`http://127.0.0.1:8000/groupe/${this.groupe.id}/`, this.groupe)
                 .then(() => {
                     alert('Groupe sauvegardé avec succès');
@@ -61,6 +66,10 @@ const groupe = {
                 .catch(error => alert('Erreur lors de la sauvegarde du groupe: ' + error.message));
         },
         confirmDelete() {
+            if (this.groupe.id === 1) {
+                alert("Favoris ne peut pas être supprimé.");
+                return;
+            }
             if (confirm("Voulez-vous vraiment supprimer ce groupe ?")) {
                 axios.delete(`http://127.0.0.1:8000/groupe/${this.groupe.id}/`)
                     .then(() => {
@@ -76,6 +85,7 @@ const groupe = {
         axios.get(`http://127.0.0.1:8000/groupe/${groupeId}`)
             .then(response => {
                 this.groupe = response.data;
+                this.originalNom = this.groupe.nom; // Sauvegarder le nom original à la montée du composant
                 this.loadContactsDuGroupe();
             })
             .catch(error => {
